+++ api/migrations/20241012140420_Notifications.js
@@ -1,13 +1,13 @@
 export const up = function (knex) {
   return knex.schema.createTable("notifications", function (table) {
-    table.increments();
+    table.bigIncrements();
     table
       .enu("type", ["Like", "Reshare", "Follow", "Reply", "Mention"])
       .notNullable();
 
-    table.integer("user_id").unsigned().notNullable();
-    table.integer("post_id").unsigned().nullable();
-    table.integer("reply_id").unsigned().nullable();
+    table.bigInteger("user_id").unsigned().notNullable();
+    table.bigInteger("post_id").unsigned().nullable();
+    table.bigInteger("reply_id").unsigned().nullable();
     table.integer("count").unsigned().notNullable().defaultTo(1);
     table.boolean("is_read").notNullable().defaultTo(false);
     table.timestamps();
